refactor: create tags component
